pyproject.toml
[project]
name = "photo-framer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
]

[dependency-groups]
dev = [
    "ty>=0.0.14",
]

[project.scripts]
photo-framer = "photo_framer.cli:main"
